Adding cast

+++ robo4j-hw-rpi/src/main/java/com/robo4j/hw/rpi/i2c/gps/SparkFunXA1110Device.java
@@ -109,7 +109,7 @@ public class SparkFunXA1110Device extends AbstractI2CDevice {
 
 	private void init() throws IOException {
 		System.out.println("Initializing device");
-		i2cDevice.write(0);
+		i2cDevice.write((byte) 0);
 		
 		int read = i2cDevice.read();
 		System.out.println("Init read: " + read);
